requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiomcache==0.8.2
aiosignal==1.3.2
alembic==1.15.1
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dataclasses-json==0.6.7
dotenv==0.9.9
et_xmlfile==2.0.0
expiring-dict==1.1.2
fastapi==0.115.11
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.40
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.6
langsmith==0.3.11
Mako==1.3.9
MarkupSafe==3.0.2
marshmallow==3.26.1
mmh3==4.1.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
openai==0.28.0
openpyxl==3.1.5
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pendulum==3.0.0
pillow==11.1.0
pinecone==6.0.1
pinecone-client==6.0.0
pinecone-notebooks==0.1.1
pinecone-plugin-interface==0.0.7
pinecone-text==0.9.0
propcache==0.3.0
psutil==7.0.0
psycopg2==2.9.10
pure_eval==0.2.3
py-automapper==2.1.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
setuptools==75.8.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.46.0
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
time-machine==2.16.0
tokenizers==0.21.0
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
types-requests==2.32.0.20250301
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
wget==3.2
yarl==1.18.3
zstandard==0.23.0
